requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.1
httpx>=0.25.2
